improved the bot to work with multiple guilds and changed file name from discord_bot to main
